Added various barlines, segno, coda and repeat brackets to stave.js.
These changes allow for all types of music notation with appropriate
barlines and repetition symbols needed for showing the flow of the music.
The TabdStave.js class was modified to allow for changing the number of
strings for instruments other than guitar (i.e. mandolin, banjo, hammered
dulcimer, etc.) Additional tests were added to the stabe test classes,
including an eight measure demo of a complete score. The "flow.html" web
page was changed to use the gonevill_all.font to make the additional
symboles available.