Fix #839. Don't shift more than 5% to the right of the previous context.
The negative-shift-limit was set to the notehead of the previous
context, so the formatter was free to use that space if it needed. This
caused (in some cases) different tick contexts to be aligned when
the were rendered.

Also add globalSoftmax option, for really complicated scores where
you want softmax to be run across all notes in all voices.