Fix Glyph.getOutlineBoundingBox.
getOutlineBoundingBox was not passing the first control point for either
qudratic or cubic Bezier curves and left the last (x, y) parameters to
addQuadraticCurve and addBezierCurve as undefined. This caused the
computed bounding boxes for individual curve pieces to be significantly
too small.

Despite this bug, the computed bounding boxes for an entire Glyph are
pretty much all correct. This is because the incorrect bounding boxes
for a curve did contain one end point, and taking the union of just the
curve end points yields glyph bounding boxes that are close enough to
using the bounding box of the entire curve.

Nevertheless, it seems prudent to actually compute the curve bounding
boxes correctly.

As a bonus, I've added some unit tests and replaced the implementation
of BoundingBoxComputation.addQuadraticCurve with one that actually
computes the bounding box directly.