feat(Factory): allow initialization with Renderer
The `renderer` field of `FactoryOptions` was changed to also be an
instance of `Renderer` via an union.

The other way of initializing a `Factory` remains unaffected.

A test-case had to be slightly modified since it didn't account for
the union I created.

This fixes #580 since an instance of `Renderer` can be created with DOM
elements and `Factory` can now be created with an instance of
`Renderer`.